src/main/java/com/hamster/baimihua/controller/HomeController.java
package com.hamster.baimihua.controller;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.config.annotation.web.builders.HttpSecurity;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
import org.springframework.stereotype.Component;
import org.springframework.stereotype.Controller;
import org.springframework.stereotype.Repository;
import org.springframework.stereotype.Service;
import org.springframework.web.bind.annotation.GetMapping;
@Component
@Controller
@Service
@Repository
public class HomeController {
    @GetMapping("/")
    public String home() throws Exception {
        return "home";
    }
}
